full

+++ dvd_rental_main/src/main/java/mn/ictgroup/intern/main/entity/Payment.java
@@ -5,6 +5,7 @@ import lombok.Setter;
 
 import javax.persistence.*;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -30,7 +31,8 @@ public class Payment implements Serializable {
     @Column(name = "rental_id", nullable = false)
     private Long rentalId;
     @Column(name = "amount", nullable = false)
-    private Long amount;
+    private BigDecimal amount;
+
     @Column(name = "payment_date")
     private Date paymentDate = new Date();
 }
